updated workflow to run all on branches, and to not cause conflicts
